ModCommands & Moved other registries to registries folder.
