--- forge/src/main/kotlin/militarymod/ModEntities.kt
@@ -1,10 +1,0 @@
-package militarymod
-
-import net.minecraft.world.entity.EntityType
-import net.minecraftforge.registries.DeferredRegister
-import net.minecraftforge.registries.ForgeRegistries
-
-
-object ModEntities {
-    val REGISTRY: DeferredRegister<EntityType<*>>? = DeferredRegister.create(ForgeRegistries.ENTITIES, MilitaryMod.MODID)
-}
