+++ forge/src/main/kotlin/militarymod/registries/ModBes.kt
@@ -1,0 +1,10 @@
+package militarymod.registries
+
+import militarymod.MilitaryMod
+import net.minecraft.world.level.block.entity.BlockEntityType
+import net.minecraftforge.registries.DeferredRegister
+import net.minecraftforge.registries.ForgeRegistries
+
+object ModBes {
+    val REGISTRY: DeferredRegister<BlockEntityType<*>>? = DeferredRegister.create(ForgeRegistries.BLOCK_ENTITIES, MilitaryMod.modId)
+}
