+++ forge/src/main/kotlin/militarymod/registries/ModItems.kt
@@ -1,11 +1,12 @@
-package militarymod
+package militarymod.registries
 
+import militarymod.MilitaryMod
 import net.minecraft.world.item.Item
 import net.minecraftforge.registries.DeferredRegister
 import net.minecraftforge.registries.ForgeRegistries
 
 object ModItems {
-    val REGISTRY: DeferredRegister<Item> = DeferredRegister.create(ForgeRegistries.ITEMS, MilitaryMod.MODID)
+    val REGISTRY: DeferredRegister<Item> = DeferredRegister.create(ForgeRegistries.ITEMS, MilitaryMod.modId)
 
     fun register() {
         ModBlocks.registerItems(REGISTRY)
